+++ CHANGELOG.md
@@ -19,7 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 - Added patcher directory to search paths for `add`ed chuck file.
 
-- Added more callback demos for global vars
+- Added improved callback demos for global vars
 
 - Changed `remove` message to able to take more than than one shred_id